requirements.txt
## opengl-related requirements
pyglet
moderngl

# ipc
protobuf
grpcio # gRPC
grpcio-tools

# general
futures
